fix(lint): address clippy warnings
$running clippy on all targets surfaces warnings in tests and\nsupporting code that were previously unchecked. fix bool-assert\ncomparison, redundant imports, map identity, let-and-return,\nmanual repeat, field reassignment with default, collapsible if,\nunnecessary cast, unused enumerate, op-ref, cloned ref and\nuseless vec issues so make lint passes with -D warnings.

+++ src/evm/coverage/shared.rs
@@ -500,7 +500,7 @@ mod tests {
 
                     // Unique revert per thread.
                     let rev_word = i as usize;
-                    let rev_bit = (i % 64) as u64;
+                    let rev_bit = i % 64;
                     contract.reverts[rev_word] = 1 << rev_bit;
                     contract.hit_reverts.push(rev_word);
